Move test-double inside sharedTest folder for reusability
